SIGN-2009/support investigation net sdk query crash/0 (#177)
* fix: parsing empty string in response header

A recent change in the public api means that queries that only return a
single page a result will have an empty link header. We were only
checking if the link header was null. The application would therefore
crash since it would try an parse an empty linkHeader.Value. We fix this
by also checking if linkHeader.Value is an empty string

* chore: update changelog for 7.0.1